fix(frontend): eliminate UI flashes on page load
- ActiveProjectContext: initialize loading=true so AppContent never
  shows NoProjectAccess before the first fetch completes
- NavigationProvider: remove mounted guard that deferred nav items to
  a second render, causing the sidebar to appear empty on first paint
- Split (protected)/layout.tsx into a server component shell that
  fetches the active project server-side (same pattern as org name)
  and ProtectedLayoutClient.tsx for the client-side providers; seeds
  ActiveProjectProvider with the initial project so the project name
  is available on first render with no flash
- Move CssBaseline from ThemeProvider into LayoutContent inside
  AppRouterCacheProvider to fix the emotion CSS injection order that
  caused a hydration mismatch (server rendered <style>, client
  expected <div>)
